# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LiturgiaFlow Pro API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  details?: string;
}

export type ReaderLevel = (typeof ReaderLevel)[keyof typeof ReaderLevel];

export const ReaderLevel = {
  Principiante: "Principiante",
  Experto: "Experto",
} as const;

export interface Reader {
  id: number;
  name: string;
  whatsapp: string;
  level: ReaderLevel;
  createdAt: string;
}

export type CreateReaderInputLevel =
  (typeof CreateReaderInputLevel)[keyof typeof CreateReaderInputLevel];

export const CreateReaderInputLevel = {
  Principiante: "Principiante",
  Experto: "Experto",
} as const;

export interface CreateReaderInput {
  /**
   * @minLength 2
   * @maxLength 100
   */
  name: string;
  /**
   * @minLength 7
   * @maxLength 20
   */
  whatsapp: string;
  level: CreateReaderInputLevel;
}

export type UpdateReaderInputLevel =
  (typeof UpdateReaderInputLevel)[keyof typeof UpdateReaderInputLevel];

export const UpdateReaderInputLevel = {
  Principiante: "Principiante",
  Experto: "Experto",
} as const;

export interface UpdateReaderInput {
  /**
   * @minLength 2
   * @maxLength 100
   */
  name?: string;
  /**
   * @minLength 7
   * @maxLength 20
   */
  whatsapp?: string;
  level?: UpdateReaderInputLevel;
}

export interface Unavailability {
  id: number;
  readerId: number;
  blockedDate: string;
  readerName?: string;
}

export interface CreateUnavailabilityInput {
  readerId: number;
  blockedDate: string;
}

export interface CalendarEntry {
  id: number;
  date: string;
  role: string;
  readerId?: number;
  readerName?: string;
  logisticComment?: string;
  isVacant: boolean;
  versionTimestamp?: string;
  liturgicalSeason?: string;
}

export type GenerateCalendarInputPeriod =
  (typeof GenerateCalendarInputPeriod)[keyof typeof GenerateCalendarInputPeriod];

export const GenerateCalendarInputPeriod = {
  "15days": "15days",
  "1month": "1month",
} as const;

export interface GenerateCalendarInput {
  startDate: string;
  period: GenerateCalendarInputPeriod;
  roles?: string[];
}

export interface UpdateCalendarEntryInput {
  readerId?: number;
  role?: string;
  logisticComment?: string;
  isVacant?: boolean;
}

export type GetUnavailabilityParams = {
  readerId?: number;
};

export type GetCalendarParams = {
  startDate?: string;
  endDate?: string;
};