# lib/api-zod/src/generated/types/calendarEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LiturgiaFlow Pro API
 * OpenAPI spec version: 0.1.0
 */

export interface CalendarEntry {
  id: number;
  date: Date;
  role: string;
  readerId?: number;
  readerName?: string;
  logisticComment?: string;
  isVacant: boolean;
  versionTimestamp?: Date;
  liturgicalSeason?: string;
}