# lib/api-zod/src/generated/types/createReaderInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LiturgiaFlow Pro API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateReaderInputLevel } from "./createReaderInputLevel";

export interface CreateReaderInput {
  /**
   * @minLength 2
   * @maxLength 100
   */
  name: string;
  /**
   * @minLength 7
   * @maxLength 20
   */
  whatsapp: string;
  level: CreateReaderInputLevel;
}