# lib/api-zod/src/generated/types/createUnavailabilityInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LiturgiaFlow Pro API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateUnavailabilityInput {
  readerId: number;
  blockedDate: Date;
}