# lib/api-zod/src/generated/types/generateCalendarInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LiturgiaFlow Pro API
 * OpenAPI spec version: 0.1.0
 */
import type { GenerateCalendarInputPeriod } from "./generateCalendarInputPeriod";

export interface GenerateCalendarInput {
  startDate: Date;
  period: GenerateCalendarInputPeriod;
  roles?: string[];
}