# lib/api-zod/src/generated/types/reader.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LiturgiaFlow Pro API
 * OpenAPI spec version: 0.1.0
 */
import type { ReaderLevel } from "./readerLevel";

export interface Reader {
  id: number;
  name: string;
  whatsapp: string;
  level: ReaderLevel;
  createdAt: Date;
}