# lib/api-zod/src/generated/types/unavailability.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LiturgiaFlow Pro API
 * OpenAPI spec version: 0.1.0
 */

export interface Unavailability {
  id: number;
  readerId: number;
  blockedDate: Date;
  readerName?: string;
}